pyproject.toml
[project]
name = "app"
version = "1.0.0"
description = "A loan api"
authors = [
    {name = "Alexandre Silva",email = "alexandresilva933@gmail.com"}
]
license = {text = "MIT License"}
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.1.7,<6.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "pytest-django (>=4.10.0,<5.0.0)",
    "flake8 (>=7.1.2,<8.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "validate-docbr (>=1.10.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = 'loans_api.settings.dev'
minversion = "6.0"
addopts = "-v -ra -q"
log_cli = true
log_format = "%(asctime)s %(levelname)s %(message)s"
log_date_format = "%Y-%m-%d %H:%M:%S"
filterwarnings = "ignore"
testpaths = [
    "tests",
]